After some Hardware Problems see OneNote
Trigger select etc. now working but programmed mode not the desired one
yet due to problems when multiple tiggers are plugged in and trying to
trigger the system.
Planing to reduce the choice of triggers and therefore giving the option
to really chose where to get the trigger signal from.