If a scan was aborted / halted TILDA will not continue  with the next repetitions anymore.
improved TimeResolvedSequencerDummy.py artificial data building
to be able to check for number of bunches.
had to modify tildaPipeline.py and tildaNodes.py for that,
 but with default value None nothing should happen

Introduced Test script to display the selected
 number of bunches per Step, e.g. [(bun0, bun5)]
 etc. this can be selected for each track

changed number of backup files for the log.

introduced timeout and warning when kepco scan is completed
according to the fpga, but not all dmm readings have ben acquired,
which is normally the indication for a complete kepco scan

Main.py
fixed init of dmms, when they are only selected in duringScan and no preScan was performed at all.