Added a DataloggerBase class that can be used as a template for any database-like datalogging. The InfluxStream class is already ported, the SQLStream class is not but is fully compatible - so the port should be fairly easy.
Fixed a bug in the Importer so that the real Configs are actually loaded.

Added an option "useinfluxoversql" in InfluxConfigDraft. When this is set to true, TILDA will import InlfuxStream instead of SQLStream in ScanMain. This provides an easy way to play with the InfluxStream class until it gets its own UI tab.